+++ src/f10_world/test/test_world_etl03_03_face_pidgins_to_event_pidgins.py
@@ -1,4 +1,4 @@
-from src.f00_instrument.file import create_path
+from src.f00_instrument.file import create_path, get_dir_filenames, create_dir
 from src.f04_gift.atom_config import face_id_str
 from src.f08_pidgin.pidgin_config import (
     event_id_str,
@@ -22,86 +22,10 @@ from pandas.testing import assert_frame_equal as pandas_testing_assert_frame_equ
 from os.path import exists as os_path_exists
 
 
-def test_WorldUnit_event_pidgins_to_pidgin_csv_files_Scenario0_3Event_road(
+def test_WorldUnit_face_pidgins_to_event_pidgins_Scenario0_road_Two_face_ids(
     env_dir_setup_cleanup,
 ):
     # ESTABLISH
-    sue_str = "Sue"
-    bob_str = "Bob"
-    zia_str = "Zia"
-    casa_otx = "fizz,casa"
-    casa_inx = "fizz,casaita"
-    clean_otx = "fizz,casa,clean"
-    clean_inx = "fizz,casaita,limpio"
-    event3 = 3
-    event7 = 7
-    event9 = 9
-    road_file_columns = [
-        face_id_str(),
-        event_id_str(),
-        otx_road_str(),
-        inx_road_str(),
-        otx_wall_str(),
-        inx_wall_str(),
-        unknown_word_str(),
-    ]
-    x_nan = float("nan")
-    e3_road0 = [bob_str, event3, casa_otx, casa_inx, x_nan, x_nan, x_nan]
-    e3_road1 = [bob_str, event3, clean_otx, clean_inx, x_nan, x_nan, x_nan]
-    e3_road_rows = [e3_road0, e3_road1]
-    e3_road_df = DataFrame(e3_road_rows, columns=road_file_columns)
-
-    e7_road0 = [sue_str, event7, casa_otx, casa_inx, x_nan, x_nan, x_nan]
-    e7_road1 = [sue_str, event7, clean_otx, clean_inx, x_nan, x_nan, x_nan]
-    e7_road_rows = [e7_road0, e7_road1]
-    e7_road_df = DataFrame(e7_road_rows, columns=road_file_columns)
-
-    e9_road0 = [zia_str, event9, casa_otx, casa_inx, x_nan, x_nan, x_nan]
-    e9_road1 = [zia_str, event9, clean_otx, clean_inx, x_nan, x_nan, x_nan]
-    e9_road_rows = [e9_road0, e9_road1]
-    e9_road_df = DataFrame(e9_road_rows, columns=road_file_columns)
-
-    fizz_world = worldunit_shop("Fizz")
-    bob_dir = create_path(fizz_world._faces_dir, bob_str)
-    sue_dir = create_path(fizz_world._faces_dir, sue_str)
-    zia_dir = create_path(fizz_world._faces_dir, zia_str)
-    event3_dir = create_path(bob_dir, event3)
-    event7_dir = create_path(sue_dir, event7)
-    event9_dir = create_path(zia_dir, event9)
-    event3_pidgin_file_path = create_path(event3_dir, "pidgin.xlsx")
-    event7_pidgin_file_path = create_path(event7_dir, "pidgin.xlsx")
-    event9_pidgin_file_path = create_path(event9_dir, "pidgin.xlsx")
-    event3_road_csv_file_path = create_path(event3_dir, "road.csv")
-    event7_road_csv_file_path = create_path(event7_dir, "road.csv")
-    event9_road_csv_file_path = create_path(event9_dir, "road.csv")
-    road_agg_str = "road_agg"
-    upsert_sheet(event3_pidgin_file_path, road_agg_str, e3_road_df)
-    upsert_sheet(event7_pidgin_file_path, road_agg_str, e7_road_df)
-    upsert_sheet(event9_pidgin_file_path, road_agg_str, e9_road_df)
-    assert sheet_exists(event3_pidgin_file_path, road_agg_str)
-    assert sheet_exists(event7_pidgin_file_path, road_agg_str)
-    assert sheet_exists(event9_pidgin_file_path, road_agg_str)
-    assert os_path_exists(event3_road_csv_file_path) is False
-    assert os_path_exists(event7_road_csv_file_path) is False
-    assert os_path_exists(event9_road_csv_file_path) is False
-
-    # WHEN
-    fizz_world.event_pidgins_to_pidgin_csv_files()
-
-    # THEN
-    assert os_path_exists(event3_road_csv_file_path)
-    assert os_path_exists(event7_road_csv_file_path)
-    assert os_path_exists(event9_road_csv_file_path)
-    gen_event3_csv_df = open_csv(event7_dir, "road.csv")
-    print(f"{gen_event3_csv_df=}")
-    pandas_testing_assert_frame_equal(gen_event3_csv_df, e7_road_df)
-
-
-def test_etl_face_pidgins_to_event_pidgins_Scenario0_road_Two_face_ids(
-    env_dir_setup_cleanup,
-):
-    # ESTABLISH
-    fizz_world = worldunit_shop("Fizz")
     sue_str = "Sue"
     zia_str = "Zia"
     casa_otx = "fizz,casa"
@@ -121,15 +45,15 @@ def test_etl_face_pidgins_to_event_pidgins_Scenario0_road_Two_face_ids(
         unknown_word_str(),
     ]
     x_nan = float("nan")
-    z3_road3 = [zia_str, event3, clean_otx, clean_inx, x_nan, x_nan, x_nan]
-    zia_road_agg_df = DataFrame([z3_road3], columns=road_file_columns)
-
     e1_road0 = [sue_str, event7, casa_otx, casa_inx, x_nan, x_nan, x_nan]
     e1_road1 = [sue_str, event7, clean_otx, clean_inx, x_nan, x_nan, x_nan]
     e1_road2 = [sue_str, event9, clean_otx, clean_inx, x_nan, x_nan, x_nan]
     e1_road_rows = [e1_road0, e1_road1, e1_road2]
     sue_road_agg_df = DataFrame(e1_road_rows, columns=road_file_columns)
+    z1_road3 = [zia_str, event3, clean_otx, clean_inx, x_nan, x_nan, x_nan]
+    zia_road_agg_df = DataFrame([z1_road3], columns=road_file_columns)
 
+    fizz_world = worldunit_shop("fizz")
     sue_dir = create_path(fizz_world._faces_dir, sue_str)
     zia_dir = create_path(fizz_world._faces_dir, zia_str)
     event3_dir = create_path(zia_dir, event3)
